requirements.txt
flask==2.3.3
flask-cors==4.0.0
bcrypt==4.0.1
PyJWT==2.8.0
google-generativeai==0.3.2
langchain==0.3.26
langchain-google-genai==2.1.5
langchain-community==0.3.26
langchain-core==0.3.66
elevenlabs==0.2.27
numpy==2.3.1
requests==2.31.0
python-dotenv==1.0.0
pytz==2023.3
SpeechRecognition==3.10.0
pyttsx3==2.90
beautifulsoup4==4.12.2
Werkzeug==2.3.7
